#56 Fix profile_left wrapper height==0 #56
Hi :)

In header.php do you by any chance have this near the top of your page:

<!DOCTYPE html>

Some students have had an issue with this before. While it is recommended by most sources to specify the doctype, it is not required. Failing to add doctype means that the browser renders in quirks mode. This is just a way that some browsers render pages so they maintain backwards compatibility with older versions of the browsers.

If you want to keep using this, you might want to look into giving your side bar of the profile a fixed position:

.profile_left {
    position: fixed;
    top: 40px; /* position just under the top nav bar */
    bottom: 0px; /* position at bottom */
    height: auto; /* height is automatically calculated based on top and bottom position */
}
You would then need to push the profile main column over to the left:

.profile_main_column {
    left: 165px;
    position: relative;
}
Alternately, you could just scroll the posts container by adding these to the .profile_main_column style:

max-height: 90%;
overflow-y: auto;
Thanks,

Reece